Refactored Deck Display +1
- Refactored Deck Display
- Changed Create Deck form for use in Edit Deck
