+++ src/Layout/DeckLayout/Deck.js
@@ -3,43 +3,10 @@ import {useParams, Link, Switch, Route} from "react-router-dom"
 import {readDeck} from "../../utils/api"
 
 import NotFound from "../NotFound"
-import CardList from "./CardList"
-import {Study, Edit, Delete} from "../Common/Buttons"
+
+import DeckDisplay from "./DeckDisplay"
 import DeckNav from "./DeckNav"
 import { useRouteMatch } from "react-router-dom"
-
-function DeckDisplay({deck, deleteFunction = () => {}})
-{
-    return (
-        <>
-            { deck ? (
-
-            <div>
-                <div>
-                    <h3>{deck.name}</h3>
-                    <p>{deck.description}</p>
-                </div>
-                <div className="d-flex justify-content-between">
-                    <div className="m-0">
-                        <Edit to="/" className="mr-2" />
-                        <Study to="/" className="mr-2" />
-                        <Link to="/" className="btn btn-primary bi-plus"> Add Cards</Link>
-                    </div>
-                    <Delete deleteFunction={deleteFunction} to="/" id={deck.id}/>
-
-                </div>
-
-                <CardList cards={deck.cards}/>
-
-            </div>
-            ) : <NotFound />
-            }
-
-
-        </>
-
-    );
-}
 
 function Deck({deleteFunction = () => {}}) {
     const [deck, setDeck] = useState({})
